files renamed or deleted
